+++ packages/ssr-data-plugin/src/constants.ts
@@ -1,3 +1,3 @@
-export const LOGGER_CONTEXT = "ssr-data-plugin";
+const LOGGER_CONTEXT = "ssr-data-plugin";
 
 export const ERROR_PREFIX = `[@real-router/${LOGGER_CONTEXT}]`;